+++ server/src/main/java/org/opensearch/cluster/service/ClusterManagerThrottlingStats.java
@@ -8,8 +8,15 @@
 
 package org.opensearch.cluster.service;
 
+import org.opensearch.common.io.stream.StreamInput;
+import org.opensearch.common.io.stream.StreamOutput;
+import org.opensearch.common.io.stream.Writeable;
 import org.opensearch.common.metrics.CounterMetric;
+import org.opensearch.common.xcontent.ToXContent;
+import org.opensearch.common.xcontent.ToXContentFragment;
+import org.opensearch.common.xcontent.XContentBuilder;
 
+import java.io.IOException;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -17,9 +24,13 @@ import java.util.concurrent.ConcurrentHashMap;
  * Contains stats of Cluster Manager Task Throttling.
  * It stores the total cumulative count of throttled tasks per task type.
  */
-public class ClusterManagerThrottlingStats implements ClusterManagerTaskThrottlerListener {
+public class ClusterManagerThrottlingStats implements ClusterManagerTaskThrottlerListener, Writeable, ToXContentFragment {
 
-    private Map<String, CounterMetric> throttledTasksCount = new ConcurrentHashMap<>();
+    private Map<String, CounterMetric> throttledTasksCount;
+
+    public ClusterManagerThrottlingStats() {
+        throttledTasksCount = new ConcurrentHashMap<>();
+    }
 
     private void incrementThrottlingCount(String type, final int counts) {
         throttledTasksCount.computeIfAbsent(type, k -> new CounterMetric()).inc(counts);
@@ -39,4 +50,38 @@ public class ClusterManagerThrottlingStats implements ClusterManagerTaskThrottle
     public void onThrottle(String type, int counts) {
         incrementThrottlingCount(type, counts);
     }
+
+    @Override
+    public void writeTo(StreamOutput out) throws IOException {
+        out.writeInt(throttledTasksCount.size());
+        for (Map.Entry<String, CounterMetric> entry : throttledTasksCount.entrySet()) {
+            out.writeString(entry.getKey());
+            out.writeLong(entry.getValue().count());
+        }
+    }
+
+    public ClusterManagerThrottlingStats(StreamInput in) throws IOException {
+        int throttledTaskEntries = in.readInt();
+        throttledTasksCount = new ConcurrentHashMap<>();
+        for (int i = 0; i < throttledTaskEntries; i++) {
+            String taskType = in.readString();
+            long throttledTaskCount = in.readLong();
+            onThrottle(taskType, (int) throttledTaskCount);
+        }
+    }
+
+    @Override
+    public XContentBuilder toXContent(XContentBuilder builder, ToXContent.Params params) throws IOException {
+        builder.startObject("cluster_manager_throttling");
+        builder.startObject("cluster_manager_stats");
+        builder.field("TotalThrottledTasks", getTotalThrottledTaskCount());
+        builder.startObject("ThrottledTasksPerTaskType");
+        for (Map.Entry<String, CounterMetric> entry : throttledTasksCount.entrySet()) {
+            builder.field(entry.getKey(), entry.getValue().count());
+        }
+        builder.endObject();
+        builder.endObject();
+        return builder.endObject();
+    }
+
 }